+++ model/commitqueue/commit_queue.go
@@ -5,8 +5,6 @@ import (
 	"time"
 
 	mgobson "github.com/evergreen-ci/evergreen/db/mgo/bson"
-	"github.com/evergreen-ci/evergreen/model/event"
-	"github.com/evergreen-ci/evergreen/model/task"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
 	"github.com/pkg/errors"
@@ -208,59 +206,4 @@ func ClearAllCommitQueues() (int, error) {
 	}
 
 	return clearedCount, nil
-}
-
-func RemoveCommitQueueItemForVersion(projectId, version string, user string) (*CommitQueueItem, error) {
-	cq, err := FindOneId(projectId)
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting commit queue for project '%s'", projectId)
-	}
-	if cq == nil {
-		return nil, errors.Errorf("no commit queue found for project '%s'", projectId)
-	}
-
-	issue := ""
-	for _, item := range cq.Queue {
-		if item.Version == version {
-			issue = item.Issue
-		}
-	}
-	if issue == "" {
-		return nil, nil
-	}
-
-	return cq.RemoveItemAndPreventMerge(issue, true, user)
-}
-
-func (cq *CommitQueue) RemoveItemAndPreventMerge(issue string, versionExists bool, user string) (*CommitQueueItem, error) {
-	removed, err := cq.Remove(issue)
-	if err != nil {
-		return removed, errors.Wrapf(err, "removing item '%s' from commit queue for project '%s'", issue, cq.ProjectID)
-	}
-
-	if removed == nil {
-		return nil, nil
-	}
-	if versionExists {
-		err = preventMergeForItem(*removed, user)
-	}
-
-	return removed, errors.Wrapf(err, "preventing merge for item '%s' in commit queue for project '%s'", issue, cq.ProjectID)
-}
-
-func preventMergeForItem(item CommitQueueItem, user string) error {
-	// Disable the merge task
-	mergeTask, err := task.FindMergeTaskForVersion(item.Version)
-	if err != nil {
-		return errors.Wrapf(err, "finding merge task for '%s'", item.Issue)
-	}
-	if mergeTask == nil {
-		return errors.New("merge task doesn't exist")
-	}
-	event.LogMergeTaskUnscheduled(mergeTask.Id, mergeTask.Execution, user)
-	if err = mergeTask.SetDisabledPriority(user); err != nil {
-		return errors.Wrap(err, "disabling merge task")
-	}
-
-	return nil
 }
