+++ model/task/task.go
@@ -1020,7 +1020,7 @@ func (t *Task) cacheExpectedDuration() error {
 // to a pod.
 func (t *Task) MarkAsContainerDispatched(ctx context.Context, env evergreen.Environment, podID, agentVersion string) error {
 	dispatchedAt := time.Now()
-	query := isContainerTaskScheduledQuery()
+	query := IsContainerTaskScheduledQuery()
 	query[StatusKey] = evergreen.TaskUndispatched
 	query[ContainerAllocatedKey] = true
 	update := bson.M{
@@ -1402,30 +1402,6 @@ func UnscheduleStaleUnderwaterHostTasks(distroID string) (int, error) {
 	}
 
 	return info.Updated, nil
-}
-
-// DisableStaleContainerTasks disables all container tasks that have been
-// scheduled to run for a long time without actually dispatching the task.
-func DisableStaleContainerTasks(caller string) error {
-	query := isContainerTaskScheduledQuery()
-	query[ActivatedTimeKey] = bson.M{"$lte": time.Now().Add(-UnschedulableThreshold)}
-
-	tasks, err := FindAll(db.Query(query))
-	if err != nil {
-		return errors.Wrap(err, "finding tasks that need to be disabled")
-	}
-
-	grip.Info(message.Fields{
-		"message":   "disabling container tasks that are still scheduled to run but are stale",
-		"num_tasks": len(tasks),
-		"caller":    caller,
-	})
-
-	if err := DisableTasks(tasks, caller); err != nil {
-		return errors.Wrap(err, "disabled stale container tasks")
-	}
-
-	return nil
 }
 
 // LegacyDeactivateStepbackTasksForProject deactivates and aborts any scheduled/running tasks
@@ -2191,98 +2167,6 @@ func (t *Task) UpdateHeartbeat() error {
 			},
 		},
 	)
-}
-
-// SetDisabledPriority sets the priority of a task so it will never run. If it's
-// a display task, it will disable the display task and all of its child
-// execution tasks. If it's an execution task, its parent display task will not
-// be updated. It also deactivates the task and any tasks that depend on it.
-func (t *Task) SetDisabledPriority(caller string) error {
-	t.Priority = evergreen.DisabledTaskPriority
-
-	ids := append([]string{t.Id}, t.ExecutionTasks...)
-	_, err := UpdateAll(
-		bson.M{IdKey: bson.M{"$in": ids}},
-		bson.M{"$set": bson.M{PriorityKey: evergreen.DisabledTaskPriority}},
-	)
-	if err != nil {
-		return errors.Wrap(err, "updating priority")
-	}
-
-	query := db.Query(bson.M{
-		IdKey: bson.M{"$in": ids},
-	}).WithFields(ExecutionKey)
-	tasks, err := FindAll(query)
-	if err != nil {
-		return errors.Wrap(err, "finding matching tasks")
-	}
-	for _, task := range tasks {
-		event.LogTaskPriority(task.Id, task.Execution, caller, evergreen.DisabledTaskPriority)
-	}
-
-	return t.DeactivateTask(caller)
-}
-
-// DisableTasks is the same as (*Task).SetDisabledPriority but for many tasks.
-func DisableTasks(tasks []Task, caller string) error {
-	if len(tasks) == 0 {
-		return nil
-	}
-
-	tasksPresent := map[string]struct{}{}
-	var taskIDs []string
-	var execTaskIDs []string
-	for _, t := range tasks {
-		tasksPresent[t.Id] = struct{}{}
-		taskIDs = append(taskIDs, t.Id)
-		execTaskIDs = append(execTaskIDs, t.ExecutionTasks...)
-	}
-
-	_, err := UpdateAll(
-		ByIds(append(taskIDs, execTaskIDs...)),
-		bson.M{"$set": bson.M{PriorityKey: evergreen.DisabledTaskPriority}},
-	)
-	if err != nil {
-		return errors.Wrap(err, "updating task priorities")
-	}
-
-	execTasks, err := findMissingTasks(execTaskIDs, tasksPresent)
-	if err != nil {
-		return errors.Wrap(err, "finding additional execution tasks")
-	}
-	tasks = append(tasks, execTasks...)
-
-	for _, t := range tasks {
-		t.Priority = evergreen.DisabledTaskPriority
-		event.LogTaskPriority(t.Id, t.Execution, caller, evergreen.DisabledTaskPriority)
-	}
-
-	if err := DeactivateTasks(tasks, true, caller); err != nil {
-		return errors.Wrap(err, "deactivating dependencies")
-	}
-
-	return nil
-}
-
-// findMissingTasks finds all tasks whose IDs are missing from tasksPresent.
-func findMissingTasks(taskIDs []string, tasksPresent map[string]struct{}) ([]Task, error) {
-	var missingTaskIDs []string
-	for _, id := range taskIDs {
-		if _, ok := tasksPresent[id]; ok {
-			continue
-		}
-		missingTaskIDs = append(missingTaskIDs, id)
-	}
-	if len(missingTaskIDs) == 0 {
-		return nil, nil
-	}
-
-	missingTasks, err := FindAll(db.Query(ByIds(missingTaskIDs)))
-	if err != nil {
-		return nil, err
-	}
-
-	return missingTasks, nil
 }
 
 // GetRecursiveDependenciesUp returns all tasks recursively depended upon
