+++ model/task_lifecycle.go
@@ -147,6 +147,91 @@ func SetActiveStateById(id, user string, active bool) error {
 	return SetActiveState(user, active, *t)
 }
 
+func DisableTasks(caller string, tasks ...task.Task) error {
+	if len(tasks) == 0 {
+		return nil
+	}
+
+	tasksPresent := map[string]struct{}{}
+	var taskIDs []string
+	var execTaskIDs []string
+	for _, t := range tasks {
+		tasksPresent[t.Id] = struct{}{}
+		taskIDs = append(taskIDs, t.Id)
+		execTaskIDs = append(execTaskIDs, t.ExecutionTasks...)
+	}
+
+	_, err := task.UpdateAll(
+		task.ByIds(append(taskIDs, execTaskIDs...)),
+		bson.M{"$set": bson.M{task.PriorityKey: evergreen.DisabledTaskPriority}},
+	)
+	if err != nil {
+		return errors.Wrap(err, "updating task priorities")
+	}
+
+	execTasks, err := findMissingTasks(execTaskIDs, tasksPresent)
+	if err != nil {
+		return errors.Wrap(err, "finding additional execution tasks")
+	}
+	tasks = append(tasks, execTasks...)
+
+	for _, t := range tasks {
+		t.Priority = evergreen.DisabledTaskPriority
+		event.LogTaskPriority(t.Id, t.Execution, caller, evergreen.DisabledTaskPriority)
+	}
+
+	if err := SetActiveState(caller, false, tasks...); err != nil {
+		return errors.Wrap(err, "deactivating dependencies")
+	}
+
+	return nil
+}
+
+// findMissingTasks finds all tasks whose IDs are missing from tasksPresent.
+func findMissingTasks(taskIDs []string, tasksPresent map[string]struct{}) ([]task.Task, error) {
+	var missingTaskIDs []string
+	for _, id := range taskIDs {
+		if _, ok := tasksPresent[id]; ok {
+			continue
+		}
+		missingTaskIDs = append(missingTaskIDs, id)
+	}
+	if len(missingTaskIDs) == 0 {
+		return nil, nil
+	}
+
+	missingTasks, err := task.FindAll(db.Query(task.ByIds(missingTaskIDs)))
+	if err != nil {
+		return nil, err
+	}
+
+	return missingTasks, nil
+}
+
+// DisableStaleContainerTasks disables all container tasks that have been
+// scheduled to run for a long time without actually dispatching the task.
+func DisableStaleContainerTasks(caller string) error {
+	query := task.IsContainerTaskScheduledQuery()
+	query[task.ActivatedTimeKey] = bson.M{"$lte": time.Now().Add(-task.UnschedulableThreshold)}
+
+	tasks, err := task.FindAll(db.Query(query))
+	if err != nil {
+		return errors.Wrap(err, "finding tasks that need to be disabled")
+	}
+
+	grip.Info(message.Fields{
+		"message":   "disabling container tasks that are still scheduled to run but are stale",
+		"num_tasks": len(tasks),
+		"caller":    caller,
+	})
+
+	if err := DisableTasks(caller, tasks...); err != nil {
+		return errors.Wrap(err, "disabled stale container tasks")
+	}
+
+	return nil
+}
+
 // activatePreviousTask will set the Active state for the first task with a
 // revision order number less than the current task's revision order number.
 // originalStepbackTask is only specified if we're first activating the generator for a generated task.
@@ -821,7 +906,7 @@ func tryDequeueAndAbortCommitQueueVersion(p *patch.Patch, cq commitqueue.CommitQ
 		"patch":   issue,
 	}))
 
-	removed, err := cq.RemoveItemAndPreventMerge(issue, true, caller)
+	removed, err := RemoveItemAndPreventMerge(&cq, issue, true, caller)
 	grip.Debug(message.Fields{
 		"message": "removing commit queue item",
 		"issue":   issue,
